tests/requirements.txt
# E2E Testing Dependencies

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0  # for parallel test execution

# HTTP client
httpx>=0.24.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Performance and metrics
psutil>=5.9.0

# Reporting
pytest-json-report>=1.5.0
pytest-cov>=4.1.0

# Optional: async support
aiohttp>=3.9.0
asyncio>=3.4.3
